[이지은] feat: 좋아요시스템 좋아요 수 출력 [HMS]

+++ jieunLee/backend/src/main/java/com/example/backend/controller/BoardController.java
@@ -33,7 +33,7 @@ public class BoardController {
     }
 
     @GetMapping("/list/{memberNo}")
-    public List<Board> memberBoardList(@PathVariable("memberNo") Integer memberNo) {
+    public List<Board> memberBoardList(@PathVariable("memberNo") Long memberNo) {
         log.info("memberBoardList()");
 
         return service.memberBoardList(memberNo);
